+++ app/Http/Controllers/RoomController.php
@@ -4,32 +4,46 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Models\Room;
+use App\Http\Resources\Room as RoomResource;
+use App\Http\Resources\RoomCollection;
 
 class RoomController extends Controller
 {
+    public static $messages = [
+        'required' => 'El campo :attribute es obligatorio.',
+    ];
+
+    public static $rules = [
+        'name' =>'required|string|unique:rooms|max:255', 
+    ];
+
     public function index(){
-        return Room::all();
+        return new RoomCollection(Room::all());
     }
 
     public function show(Room $room){
-        return $room;
+        return response()->json(new RoomResource($room), 200);
     }
 
     public function store (Request $request){
+        $this->authorize('create',Room::class);
+        $request->validate( self::$rules, self::$messages);
         $room = Room::create($request->all());
-
         return response()->json($room, 201);
     }
 
     public function update (Request $request, Room $room){
+        $this->authorize('update',$room);
+        $request->validate([
+            'name' =>'required|string|unique:rooms,name,'.$room->id.'|max:255', 
+        ],self::$messages);
         $room ->update($request->all());
-        
         return response()->json($room, 200);
     }
 
-    public function delete (Room $room ){
+    public function delete (Room $room ){ 
+        $this->authorize('delete',$room);
         $room->delete();
-
         return response()->json(null, 204);
     }
 }
